dw2102: check initialization replies
In tt_s2_4600_frontend_attach() initialization commands are sent over
USB to the device.  The replies were not checked for success.

Now it is checked, whether the proper values are returned.

Also, the error messages are more detailed now.  Before, the same
message was used with different initialization commands, so it was
impossible to tell which command failed.

Signed-off-by: Michael Bunk <micha@freedict.org>